Updated generators (including base), README, and pylintrc.
Changed the return type of output in the base generator so a list of tuples containing the file(s) name and contents are returned. This means the generators only need to worry about using strings to build valid programs and not writing them to disk. All I/O is now confined to the CLI program. This allows more care to be taken with the file system by exposing a force overwrite flag which will avoid unnecessary/inadvertent overwrites. Other additional flags were also added to the CLI program for verbosity and timeing cabilities. All documentation and children generators have been updated accordingly. The pylintrc file was also updated to ignore a couple message types.