Separating the scanner/parser tests from implementation.
Separating the scanner and parser tests into there own files under a 'tests' directory.
Adding a pytest config (INI) file for repeatable pytest executions.
Ensuring the pytest config as well as the pylintrc are included in the package distribution and bumping up the verion.
Also updated the README with instruction on running the package tests either through CLI or pytest.